# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Garage SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: getVehicles
      tags: [vehicles]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vehicles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"

    post:
      operationId: createVehicle
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleDetail"

    put:
      operationId: updateVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}/oil-records:
    get:
      operationId: getVehicleOilRecords
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Oil change records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OilRecord"

    post:
      operationId: createOilRecord
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OilRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OilRecord"

  # ── WORK ORDERS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, customerId, plateNumber, make, model, year]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: ["string", "null"]
        bodyType:
          type: string
        fuelType:
          type: string
        transmission:
          type: string
        currentMileage:
          type: ["integer", "null"]
        engineSize:
          type: ["string", "null"]
        vin:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    VehicleDetail:
      allOf:
        - $ref: "#/components/schemas/Vehicle"
        - type: object
          properties:
            lastOilChange:
              $ref: "#/components/schemas/OilRecord"
            nextOilChangeMileage:
              type: ["integer", "null"]
            nextOilChangeDate:
              type: ["string", "null"]
            oilChangeOverdue:
              type: boolean
            maintenanceHistory:
              type: array
              items:
                $ref: "#/components/schemas/Order"
...
    VehicleInput:
      type: object
      required: [customerId, plateNumber, make, model, year]
      properties:
        customerId:
          type: integer
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        bodyType:
          type: string
        fuelType:
          type: string
        transmission:
          type: string
        currentMileage:
          type: integer
        engineSize:
          type: string
        vin:
          type: string
        notes:
          type: string
    VehicleUpdate:
      type: object
      properties:
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        bodyType:
          type: string
        fuelType:
          type: string
        transmission:
          type: string
        currentMileage:
          type: integer
        engineSize:
          type: string
        notes:
          type: string
    OilRecord:
      type: object
      required: [id, vehicleId, brand, mileageAtChange]
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        type:
          type: string
        brand:
          type: string
        viscosity:
          type: ["string", "null"]
        quantityLiters:
          type: ["number", "null"]
        mileageAtChange:
          type: integer
        nextChangeMileage:
          type: ["integer", "null"]
        nextChangeDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    OilRecordInput:
      type: object
      required: [brand, mileageAtChange]
      properties:
        type:
          type: string
        brand:
          type: string
        viscosity:
          type: string
        quantityLiters:
          type: number
        mileageAtChange:
          type: integer
        nextChangeMileage:
          type: integer
        nextChangeDate:
          type: string
        notes:
          type: string

    # ── ORDER ──────────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, orderNumber, vehicleId, customerId, status]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        vehicleId:
          type: integer
        customerId:
          type: integer
        status:
          type: string
        priority:
          type: string
        customerName:
          type: string
        vehiclePlate:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        estimatedCostMin:
          type: ["number", "null"]
        estimatedCostMax:
          type: ["number", "null"]
        grandTotal:
          type: ["number", "null"]
        receivedAt:
          type: string
        estimatedDeliveryAt:
          type: ["string", "null"]
        actualDeliveryAt:
          type: ["string", "null"]
        assignedTechnicianId:
          type: ["integer", "null"]
        assignedTechnicianName:
          type: ["string", "null"]
        paymentStatus:
          type: string
        customerComplaint:
          type: ["string", "null"]
        createdAt:
          type: string
...