# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Garage SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleUpdate {
  plateNumber?: string;
  make?: string;
  model?: string;
  year?: number;
  color?: string;
  bodyType?: string;
  fuelType?: string;
  transmission?: string;
  currentMileage?: number;
  engineSize?: string;
  notes?: string;
}